+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Removed the default values for `searchFillColor` and `searchCursorColor` (now defaults to `null`).
+
 ## 0.0.1
 
 * Initial creation of `ff_drop_down_list` library, forked from `drop_down_list` library [(Mindinventory/drop_down_list)](https://github.com/Mindinventory/drop_down_list).